Fixing view tests and moving memory_excel function into the test data file
